% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asgene.R
\name{Asgene}
\alias{Asgene}
\title{AsgeneDB: A curated orthology arsenic metabolism gene database for metagenome annotation
}
\usage{
Asgene(
  analysis = "",
  workdir = "",
  method = "",
  toolpath = "",
  search_parameters = "",
  seqtype = "",
  filetype = "",
  PE = TRUE,
  out = ""
)
}
\arguments{
\item{analysis}{Choose the target you want to analyze,abundance or taxonomy; type = "character"; default = "abundance"}

\item{workdir}{Specify directory for sequence file location;type = "character"; default = "./"}

\item{method}{Specify the database searching tool you plan to use, currently diamond, usearch and blast are supported; type = "character"; default = "diamond"}

\item{toolpath}{Specify directory for searching tool location; type = "character"; default = "./"}

\item{search_parameters}{Define metagenomic comparison parameters; type = "character"; default = "-e 1e-4 -p 28 --query-cover 80 --id 50"}

\item{seqtype}{Specify your sequence type, nucl or prot; type = "character"; default = "nucl"}

\item{filetype}{Specify the extensions of your sequence files, e.g. fastq, fastq.gz, fasta,fasta.gz, fq, fq.gz, fa, fa.gz; type = "character"; default = "fasta"}

\item{PE}{Specifies whether your metagenomic data were PE files; type = "logical"; default = TRUE}

\item{out}{Specify the directory for the final result file; type = "character"; default = "./"}
}
\value{

}
\description{
A manually curated arsenic functional gene database/tool (AsgeneDB) was developed for rapid and accurate metagenomic analysis.
}
\examples{
library(Asgene)
Asgene(analysis = "abundance", workdir = "./", method = "diamond", toolpath = "./", search_parameters = "-e 1e-4 -p 28 --query-cover 80 --id 50",seqtype = "nucl", filetype = "fasta", PE="TRUE", out = "./")
